+++ _distribution_random_float.d.ts
@@ -1,0 +1,2 @@
+export declare function _distribution_random_float(min: number | undefined, max: number | undefined, _distribution: any): number;
+export { _distribution_random_float as _float__random__distribution };