+++ _distribution_random_int.d.ts
@@ -1,0 +1,2 @@
+export declare function _distribution_random_int(min: number | undefined, max: number | undefined, $distribution: any): number;
+export { _distribution_random_int as _int__random__distribution };