+++ _normal_random.d.ts
@@ -1,0 +1,2 @@
+export declare function _normal_random(): number;
+export { _normal_random as _random__normal };