+++ _normal_random_float.d.ts
@@ -1,0 +1,2 @@
+export declare function _normal_random_float(min?: number, max?: number): number;
+export { _normal_random_float as _float__random__normal };