+++ _normal_random_int.d.ts
@@ -1,0 +1,2 @@
+export declare function _normal_random_int(min?: number, max?: number): number;
+export { _normal_random_int as _int__random__normal };