+++ _uniform_random.d.ts
@@ -1,0 +1,2 @@
+export declare function _uniform_random(): number;
+export { _uniform_random as _random__uniform };