+++ _uniform_random.ts
@@ -1,0 +1,6 @@
+export function _uniform_random() {
+	return Math.random()
+}
+export {
+	_uniform_random as _random__uniform
+}