+++ _uniform_random_float.d.ts
@@ -1,0 +1,2 @@
+export declare function _uniform_random_float(min?: number, max?: number): number;
+export { _uniform_random_float as _float__random__uniform };