+++ _uniform_random_int.d.ts
@@ -1,0 +1,2 @@
+export declare function _uniform_random_int(min?: number, max?: number): number;
+export { _uniform_random_int as _int__random__uniform };